+++ src/libc/ft_memset.c
@@ -1,6 +1,8 @@
-void    *ft_memset(void *b, int c, unsigned int len)
+#include <stddef.h>
+
+void    *ft_memset(void *b, int c, size_t len)
 {
-    unsigned int    i;
+    size_t   i;
     unsigned char   *ptr;
 
     ptr = (unsigned char *)b;
