+++ src/libc/ft_strncmp.c
@@ -1,7 +1,8 @@
+#include <stddef.h>
 
-int ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    unsigned int i;
+    size_t  i;
 
     if ( n == 0)
         return (0);
